move expandable card to `sharedlib`
