+++ toolkit/popup/src/main/java/com/arcgismaps/toolkit/popup/internal/element/textelement/TextPopupElement.kt
@@ -24,7 +24,7 @@ import android.webkit.WebViewClient
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.viewinterop.AndroidView
-import com.arcgismaps.toolkit.popup.internal.ui.ExpandableCard
+import com.arcgismaps.toolkit.ui.ExpandableCard
 
 /**
  * Renders the following [content] as HTML.
