+++ toolkit/sharedlib/src/main/java/com/arcgismaps/toolkit/ui/ExpandableCard.kt
@@ -1,20 +1,4 @@
-/*
- * Copyright 2024 Esri
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *  http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- */
-
-package com.arcgismaps.toolkit.popup.internal.ui
+package com.arcgismaps.toolkit.ui
 
 import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.animation.Crossfade
@@ -40,12 +24,10 @@ import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.arcgismaps.toolkit.popup.R
 
 /**
  * Composable Card that has the ability to expand and collapse its [content].
@@ -53,7 +35,7 @@ import com.arcgismaps.toolkit.popup.R
  * @since 200.5.0
  */
 @Composable
-internal fun ExpandableCard(
+public fun ExpandableCard(
     modifier: Modifier = Modifier,
     title: String = "",
     description: String = "",
@@ -76,7 +58,7 @@ internal fun ExpandableCard(
             .padding(shapes.padding)
     ) {
         Column {
-            ExpandableHeader(
+            ExpandableHeader2(
                 title = title,
                 description = description,
                 expandable = toggleable,
@@ -97,7 +79,7 @@ internal fun ExpandableCard(
 }
 
 @Composable
-private fun ExpandableHeader(
+private fun ExpandableHeader2(
     title: String = "",
     description: String = "",
     expandable: Boolean,
@@ -150,7 +132,7 @@ private fun ExpandableHeader(
                     modifier = Modifier
                         .padding(16.dp),
                     imageVector = if (it) Icons.Rounded.ExpandLess else Icons.Rounded.ExpandMore,
-                    contentDescription = stringResource(R.string.show_or_hide_popup_element_content)
+                    contentDescription = "Expand"
                 )
             }
         }
@@ -160,7 +142,7 @@ private fun ExpandableHeader(
 @Preview
 @Composable
 internal fun ExpandableHeaderPreview() {
-    ExpandableHeader(
+    ExpandableHeader2(
         title = "The Title",
         colors = ExpandableCardDefaults.colors(),
         description = "the description",
@@ -183,4 +165,12 @@ private fun ExpandableCardPreview() {
         )
     }
 }
+
+internal fun Modifier.applyIf(condition: Boolean, then: Modifier.() -> Modifier): Modifier =
+    if (condition) {
+        then()
+    } else {
+        this
+    }
+
