+++ toolkit/utilitynetworks/build.gradle.kts
@@ -76,6 +76,7 @@ android {
 
 dependencies {
     api(arcgis.mapsSdk)
+    implementation(project(":sharedlib"))
     implementation(platform(libs.androidx.compose.bom))
     implementation(libs.bundles.composeCore)
     implementation(libs.bundles.core)
